dissolve interfaces
