--- internal/interfaces/token_store.go
@@ -1,24 +1,0 @@
-// Copyright (c) 2026 Lateralus Labs, LLC.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package interfaces
-
-import "context"
-
-// TokenStore defines the interface for token persistence used by Sentinel.
-// This shared interface prevents drift between storage and sentinel packages.
-type TokenStore interface {
-	KVSet(ctx context.Context, key, value string, ttlSeconds int) error
-	KVGet(ctx context.Context, key string) (string, error)
-	KVScanPrefix(ctx context.Context, prefix string) (map[string]string, error)
-}
