+++ internal/services/mcp/gateway.go
@@ -36,12 +36,11 @@ import (
 	"time"
 
 	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/interfaces"
 	"github.com/g8e-ai/g8e/internal/models"
+	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	"github.com/g8e-ai/g8e/internal/response"
 	"github.com/g8e-ai/g8e/internal/services/governance"
 	"github.com/g8e-ai/g8e/internal/services/scrubbing"
-	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	govpkg "github.com/g8e-ai/g8e/pkg/governance"
 	commonv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/common/v1"
 	operatorv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/operator/v1"
@@ -95,7 +94,7 @@ type GatewayService struct {
 	downstreamURL     string
 	a2aDownstreamURL  string
 	publicBaseURL     string
-	suspendedStore    interfaces.SuspendedTransactionStore
+	suspendedStore    storage.SuspendedTransactionStore
 	fieldPathRegistry *FieldPathRegistry
 	dbService         FieldReader
 	sessionValidator  SessionValidator
@@ -137,7 +136,7 @@ type AuditLogger interface {
 type Dependencies struct {
 	Logger           *slog.Logger
 	Responder        *response.Writer
-	SuspendedStore   interfaces.SuspendedTransactionStore
+	SuspendedStore   storage.SuspendedTransactionStore
 	ScrubbingService *scrubbing.ScrubbingService
 	MaxPayloadBytes  int64
 	Posture          string // Gateway posture: doctrine, consensus, or notary
