+++ internal/services/mcp/gateway_test.go
@@ -30,11 +30,10 @@ import (
 	"time"
 
 	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/interfaces"
 	"github.com/g8e-ai/g8e/internal/models"
+	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	"github.com/g8e-ai/g8e/internal/response"
 	"github.com/g8e-ai/g8e/internal/services/governance"
-	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	"github.com/g8e-ai/g8e/internal/services/storage/storagetest"
 	commonv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/common/v1"
 	operatorv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/operator/v1"
@@ -1837,7 +1836,7 @@ func withEnvProc(proc governance.EnvelopeProcessor) testGatewayOption {
 }
 
 // withSuspendedStore sets a custom suspended store for the test GatewayService
-func withSuspendedStore(store interfaces.SuspendedTransactionStore) testGatewayOption {
+func withSuspendedStore(store storage.SuspendedTransactionStore) testGatewayOption {
 	return func(g *GatewayService) {
 		g.suspendedStore = store
 	}
