+++ internal/services/pubsub/history_service.go
@@ -21,11 +21,10 @@ import (
 
 	"github.com/g8e-ai/g8e/internal/config"
 	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/interfaces"
 	"github.com/g8e-ai/g8e/internal/models"
+	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	"github.com/g8e-ai/g8e/internal/services/scrubbing"
 	"github.com/g8e-ai/g8e/internal/services/sqliteutil"
-	storage "github.com/g8e-ai/g8e/internal/services/storage"
 	operatorv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/operator/v1"
 	"google.golang.org/protobuf/proto"
 )
@@ -35,7 +34,7 @@ type HistoryService struct {
 	config         *config.Config
 	logger         *slog.Logger
 	client         PubSubClient
-	executionVault interfaces.ExecutionVault
+	executionVault storage.ExecutionVault
 	historyHandler *storage.HistoryHandler
 	auditStore     AuditEventRecorder // *storage.SQLAuditStore - optional for observed-state content evidence
 	scrubbing      *scrubbing.ScrubbingService
